+++ lime-runtime/src/main/java/com/here/gluecodium/model/lime/LimeBasicType.kt
@@ -54,4 +54,7 @@ class LimeBasicType(val typeId: TypeId) : LimeType(path = LimePath.EMPTY_PATH) {
 
     override val escapedName
         get() = name
+
+    override val childTypes: List<LimeTypeRef>
+        get() = if (typeId == TypeId.BLOB) listOf(LimeBasicTypeRef(TypeId.UINT8)) else emptyList()
 }